refactor: Update FAQ and QnA ID type from number to string
