+++ app/(help)/help/page.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import QuestionDropdown from "@/(club)/club/help/components/questionDropdown";
 import ClubNoticeHeader from "@/(club)/club/management/activity/notice/components/clubNoticeHeader";
 import { FAQ_DATA } from "@/data/faq";
-import { profileType } from "@/models/member";
+import { profileType } from "@/types/member";
 import { PROFILE_TYPES } from "@/data/profileType";
 import { NOTICE_DATA } from "@/data/clubNotice";
 import ClubNoticeDropdown from "@/components/dropdown/clubNoticeDropdown";
@@ -16,7 +16,7 @@ const getProfileByIndex = (index: number): profileType => {
 };
 
 const HelpPage = () => {
-  const [selectedFaq, setSelectedFaq] = useState<number | null>(null);
+  const [selectedFaq, setSelectedFaq] = useState<string | null>(null);
   const [openDropdownId, setOpenDropdownId] = useState<string | null>(null);
 
   return (
