+++ app/data/faq.ts
@@ -2,181 +2,181 @@ import { ClubFaqData } from "@/types/club";
 
 export const FAQ_DATA: ClubFaqData[] = [
   {
-    id: 1,
+    id: "1",
     title: "회원 가입은 어떻게 하나요?",
     body: "회원 가입은 홈페이지 상단의 회원 가입 버튼을 클릭하여 진행할 수 있습니다.",
     clubFaqClassification: "가입",
     clubFaqColorType: "C_TOKEN_1",
     clubMemberData: {
-      id: 1,
+      id: "1",
       name: "변성은",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 101,
+        id: "101",
         nickname: "길동이",
       },
     },
   },
   {
-    id: 2,
+    id: "2",
     title: "비밀번호를 잊어버렸어요.",
     body: "비밀번호를 잊으셨을 경우, 로그인 페이지에서 '비밀번호 찾기'를 클릭하세요.",
     clubFaqClassification: "계정",
     clubFaqColorType: "C_TOKEN_3",
     clubMemberData: {
-      id: 2,
+      id: "2",
       name: "권수연",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 102,
+        id: "102",
         nickname: "영희",
       },
     },
   },
   {
-    id: 3,
+    id: "3",
     title: "회원 탈퇴는 어떻게 하나요?",
     body: "회원 탈퇴는 '내 정보' 페이지 하단의 '회원 탈퇴' 버튼을 통해 진행할 수 있습니다.",
     clubFaqClassification: "최대5글자",
     clubFaqColorType: "C_TOKEN_2",
     clubMemberData: {
-      id: 3,
+      id: "3",
       name: "박찬진",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 103,
+        id: "103",
         nickname: "철수",
       },
     },
   },
   {
-    id: 4,
+    id: "4",
     title: "온라인인가요? 오프라인인가요?",
     body: "격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.",
     clubFaqClassification: "운영",
     clubFaqColorType: "C_TOKEN_4",
     clubMemberData: {
-      id: 3,
+      id: "3",
       name: "박찬진",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 103,
+        id: "103",
         nickname: "철수",
       },
     },
   },
   {
-    id: 5,
+    id: "5",
     title: "총 몇명 뽑나요?",
     body: "50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. ",
     clubFaqClassification: "인원",
     clubFaqColorType: "C_SUB2",
     clubMemberData: {
-      id: 3,
+      id: "3",
       name: "박찬진",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 103,
+        id: "103",
         nickname: "철수",
       },
     },
   },
   {
-    id: 6,
+    id: "6",
     title: "회원 가입은 어떻게 하나요?",
     body: "회원 가입은 홈페이지 상단의 회원 가입 버튼을 클릭하여 진행할 수 있습니다.",
     clubFaqClassification: "최대5자",
     clubFaqColorType: "C_TOKEN_1",
     clubMemberData: {
-      id: 1,
+      id: "1",
       name: "변성은",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 101,
+        id: "101",
         nickname: "길동이",
       },
     },
   },
   {
-    id: 7,
+    id: "7",
     title: "비밀번호를 잊어버렸어요.",
     body: "비밀번호를 잊으셨을 경우, 로그인 페이지에서 '비밀번호 찾기'를 클릭하세요.",
     clubFaqClassification: "계정",
     clubFaqColorType: "C_TOKEN_3",
     clubMemberData: {
-      id: 2,
+      id: "2",
       name: "권수연",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 102,
+        id: "102",
         nickname: "영희",
       },
     },
   },
   {
-    id: 8,
+    id: "8",
     title: "회원 탈퇴는 어떻게 하나요?",
     body: "회원 탈퇴는 '내 정보' 페이지 하단의 '회원 탈퇴' 버튼을 통해 진행할 수 있습니다.",
     clubFaqClassification: "최대5글자",
     clubFaqColorType: "C_TOKEN_2",
     clubMemberData: {
-      id: 3,
+      id: "3",
       name: "박찬진",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 103,
+        id: "103",
         nickname: "철수",
       },
     },
   },
   {
-    id: 9,
+    id: "9",
     title: "온라인인가요? 오프라인인가요?",
     body: "격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.격주로 진행됩니다.",
     clubFaqClassification: "운영",
     clubFaqColorType: "C_TOKEN_4",
     clubMemberData: {
-      id: 3,
+      id: "3",
       name: "박찬진",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 103,
+        id: "103",
         nickname: "철수",
       },
     },
   },
   {
-    id: 10,
+    id: "10",
     title: "총 몇명 뽑나요?",
     body: "50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. 50명 뽑습니다. ",
     clubFaqClassification: "인원",
     clubFaqColorType: "C_PRIMARY",
     clubMemberData: {
-      id: 3,
+      id: "3",
       name: "박찬진",
       clubMemberRoleType: "GENERAL",
       clubMemberStatusType: "ACTIVE",
       profileType: "ARIARI_MOUSE",
       memberData: {
-        id: 103,
+        id: "103",
         nickname: "철수",
       },
     },
